fix(schema-generator): see through type-only wrappers nested inside a sign
`evaluateExpression()` strips parentheses and the type-only assertion forms off
the outside of a value, but it cannot reach one nested in the middle: by the
time `-(1 as number)` is being read as a sign applied to something, the `as` is
inside the operand. `numberFromExpression()` already recursed through
parentheses for exactly this reason; it now recurses through `as`, `satisfies`
and `<T>x` as well.

Raised by @mathpirate in review on #4887, and taken here rather than as a
follow-up: it is the same recursive-wrapper case, a few characters from the one
already fixed.

The fixture covering wrappers is renamed `wrapped-options`, since it no longer
covers only parentheses, and gains the nested case plus a `satisfies` inside an
array element.
