+++ packages/ts-transformers/test/fixtures/schema-transform/numeric-option-defaults.expected.jsx
@@ -60,7 +60,7 @@ const configSchema = __cfHelpers.__cf_data({
 // Verifies: the toSchema options object carries signed and non-finite numbers
 // through to the emitted schema. Recognizing only bare NumericLiteral drops
 // each of these properties silently — a `-1` sentinel default just vanishes.
-// (Parentheses across all option types: see `parenthesized-options`.)
+// (Wrappers across all option types: see `wrapped-options`.)
 export { configSchema };
 // @ts-ignore: Internals
 function h(...args: any[]) { return __cfHelpers.h.apply(null, args); }
